pyproject.toml
[project]
name = "paleonews"
version = "0.1.0"
description = "Paleontology news aggregator: RSS collection, filtering, summarization, and delivery"
requires-python = ">=3.11"

dependencies = [
    "feedparser",
    "httpx",
    "anthropic",
    "python-telegram-bot",
    "pyyaml",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]

[tool.setuptools.packages.find]
include = ["paleonews*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
